Add extension to unzip the relevant binary upon gem installation
This will run the code to detect and unzip the relevant binary when the
gem is installed, so that we won't have to unzip it at runtime the first
time `wkhtmltopdf` is invoked. This is particularly useful because it
avoids potential concurrency issues when running multiple threads that
use `wkhtmltopdf` at the same time, which could result in
`Text file busy (Errno::ETXTBSY)` errors if the threads simultaneously
attempt to unzip the wkhtmltopdf binary. (If we ensure the binary is
unzipped at gem installation time, the threads won't have to unzip it
at runtime, effectively bypassing this issue.)